Reload license from DB for all cluster app servers
