+++ cmd/platform/init.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 
-	"github.com/mattermost/platform/api"
 	"github.com/mattermost/platform/app"
 	"github.com/mattermost/platform/model"
 	"github.com/mattermost/platform/utils"
@@ -41,6 +40,6 @@ func initDBCommandContext(configFileLocation string) {
 	app.NewServer()
 	app.InitStores()
 	if model.BuildEnterpriseReady == "true" {
-		api.LoadLicense()
+		app.LoadLicense()
 	}
 }
